+++ src/legacy_setup/swaps/swapHandlers/SwapHandlerBase.sol
@@ -8,12 +8,12 @@ import "./Utils.sol";
 
 /// @notice Base contract for swap handlers
 abstract contract SwapHandlerBase is ISwapHandler {
-    function trySafeApprove(address token, address to, uint value) internal returns (bool, bytes memory) {
+    function trySafeApprove(address token, address to, uint256 value) internal returns (bool, bytes memory) {
         (bool success, bytes memory data) = token.call(abi.encodeWithSelector(IERC20.approve.selector, to, value));
         return (success && (data.length == 0 || abi.decode(data, (bool))), data);
     }
 
-    function safeApproveWithRetry(address token, address to, uint value) internal {
+    function safeApproveWithRetry(address token, address to, uint256 value) internal {
         (bool success, bytes memory data) = trySafeApprove(token, to, value);
 
         // some tokens, like USDT, require the allowance to be set to 0 first
@@ -28,13 +28,13 @@ abstract contract SwapHandlerBase is ISwapHandler {
     }
 
     function transferBack(address token) internal {
-        uint balance = IERC20(token).balanceOf(address(this));
+        uint256 balance = IERC20(token).balanceOf(address(this));
         if (balance > 0) Utils.safeTransfer(token, msg.sender, balance);
     }
 
-    function setMaxAllowance(address token, uint minAllowance, address spender) internal {
-        uint allowance = IERC20(token).allowance(address(this), spender);
-        if (allowance < minAllowance) safeApproveWithRetry(token, spender, type(uint).max);
+    function setMaxAllowance(address token, uint256 minAllowance, address spender) internal {
+        uint256 allowance = IERC20(token).allowance(address(this), spender);
+        if (allowance < minAllowance) safeApproveWithRetry(token, spender, type(uint256).max);
     }
 
     function revertBytes(bytes memory errMsg) internal pure {
