+++ src/legacy_setup/test/IRMFixed.sol
@@ -6,7 +6,7 @@ import "../../InterestRateModels/IIRM.sol";
 
 contract IRMFixed is IIRM {
     function computeInterestRate(address, uint256, uint256) public pure returns (uint256) {
-        return uint(1e27 * 0.1) / (86400 * 365); // not SECONDS_PER_YEAR to avoid breaking tests
+        return uint256(1e27 * 0.1) / (86400 * 365); // not SECONDS_PER_YEAR to avoid breaking tests
     }
 
     function computeInterestRateView(address vault, uint256 cash, uint256 borrows) external pure returns (uint256) {
